+++ source/DocumentValidation/JsonSchemaProvider.cs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Threading.Tasks;
 using Json.Schema;
 
-namespace Infrastructure.IncomingMessages.SchemaStore;
+namespace DocumentValidation;
 
 public class JsonSchemaProvider : SchemaProvider
 {
