+++ source/DocumentValidation/SchemaBase.cs
@@ -12,11 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace Infrastructure.IncomingMessages.SchemaStore;
+namespace DocumentValidation;
 
 public abstract class SchemaBase
 {
