+++ source/DocumentValidation/Schemas/Json/Request-Change-of-Supplier-assembly-model.schema.json
