+++ source/DocumentValidation/Schemas/Json/urn-entsoe-eu-local-extension-types.schema.json
