+++ source/DocumentValidation/Schemas/Json/urn-entsoe-eu-wgedi-codelists.schema.json
