+++ source/DocumentValidation/Schemas/Xml/urn-ediel-org-measure-notifyaggregatedmeasuredata-0-1.xsd
