+++ source/DocumentValidation/Schemas/Xml/urn-ediel-org-structure-accountingpointcharacteristics-0-1.xsd
