+++ source/DocumentValidation/Schemas/Xml/urn-ediel-org-structure-characteristicsofacustomeratanap-0-1.xsd
