+++ source/DocumentValidation/Schemas/Xml/urn-ediel-org-structure-confirmrequestchangeaccountingpointcharacteristics-0-1.xsd
