+++ source/DocumentValidation/Schemas/Xml/urn-ediel-org-structure-rejectrequestchangeaccountingpointcharacteristics-0-1.xsd
