+++ source/DocumentValidation/Schemas/Xml/urn-ediel-org-structure-requestchangeaccountingpointcharacteristics-0-1.xsd
