+++ source/DocumentValidation/Schemas/Xml/urn-ediel-org-structure-requestchangecustomercharacteristics-0-1.xsd
