+++ source/DocumentValidation/Schemas/Xml/urn-entsoe-eu-local-extension-types.xsd
