+++ source/DocumentValidation/Schemas/Xml/urn-entsoe-eu-wgedi-codelists.xsd
