+++ source/DocumentValidation/XmlSchemaProvider.cs
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Schema;
 
-namespace Infrastructure.IncomingMessages.SchemaStore;
+namespace DocumentValidation;
 
 public class XmlSchemaProvider : SchemaProvider
 {
