+++ source/Infrastructure/Configuration/IncomingMessageParsingServices.cs
@@ -18,9 +18,9 @@ using CimMessageAdapter.Messages;
 using CimMessageAdapter.Messages.RequestChangeCustomerCharacteristics;
 using CimMessageAdapter.Messages.RequestChangeOfSupplier;
 using CimMessageAdapter.Response;
+using DocumentValidation;
 using Infrastructure.IncomingMessages.RequestChangeOfSupplier;
 using Infrastructure.IncomingMessages.Response;
-using Infrastructure.IncomingMessages.SchemaStore;
 using Microsoft.Extensions.DependencyInjection;
 using MarketActivityRecord = Application.IncomingMessages.RequestChangeOfSupplier.MarketActivityRecord;
 using MessageParser = CimMessageAdapter.Messages.RequestChangeOfSupplier.MessageParser;
