+++ source/Infrastructure/IncomingMessages/RequestChangeOfSupplier/JsonMessageParser.cs
@@ -21,8 +21,8 @@ using System.Threading.Tasks;
 using Application.IncomingMessages.RequestChangeOfSupplier;
 using CimMessageAdapter.Errors;
 using CimMessageAdapter.Messages;
+using DocumentValidation;
 using Domain.OutgoingMessages;
-using Infrastructure.IncomingMessages.SchemaStore;
 using Json.Schema;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
