+++ source/Infrastructure/Infrastructure.csproj
@@ -40,69 +40,10 @@ limitations under the License.
   </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\DocumentValidation\DocumentValidation.csproj" />
       <ProjectReference Include="..\EnergySupplying.Contracts\EnergySupplying.Contracts.csproj" />
       <ProjectReference Include="..\CimMessageAdapter\CimMessageAdapter.csproj" />
       <ProjectReference Include="..\Application\Application.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Update="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-requestchangeaccountingpointcharacteristics-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </None>
-    </ItemGroup>
-
-    <ItemGroup>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Json\Confirm-request-Change-of-Supplier-assembly-model.schema.json">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Json\Reject-request-Change-of-Supplier-assembly-model.schema.json">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Json\Request-Change-of-Supplier-assembly-model.schema.json">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Json\urn-entsoe-eu-local-extension-types.schema.json">
-        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Json\urn-entsoe-eu-wgedi-codelists.schema.json">
-        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-measure-notifyaggregatedmeasuredata-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-accountingpointcharacteristics-0-1.xsd">
-        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-characteristicsofacustomeratanap-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-confirmrequestchangeaccountingpointcharacteristics-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-confirmrequestchangeofsupplier-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-genericnotification-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-rejectrequestchangeaccountingpointcharacteristics-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-rejectrequestchangeofsupplier-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-requestchangecustomercharacteristics-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-ediel-org-structure-requestchangeofsupplier-0-1.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-entsoe-eu-local-extension-types.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
-      <EmbeddedResource Include="IncomingMessages\SchemaStore\Schemas\Xml\urn-entsoe-eu-wgedi-codelists.xsd">
-        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      </EmbeddedResource>
     </ItemGroup>
 
 </Project>
