+++ source/Tests/Infrastructure/OutgoingMessages/CharacteristicsOfACustomerAtAnAP/CharacteristicsOfACustomerAtAnApDocumentWriterTests.cs
@@ -21,12 +21,12 @@ using System.Xml.Schema;
 using Application.Configuration;
 using Application.OutgoingMessages;
 using Application.OutgoingMessages.Common;
+using DocumentValidation;
 using Domain.Actors;
 using Domain.OutgoingMessages;
 using Domain.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Infrastructure.Configuration;
 using Infrastructure.Configuration.Serialization;
-using Infrastructure.IncomingMessages.SchemaStore;
 using Infrastructure.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
 using Infrastructure.OutgoingMessages.Common;
 using NodaTime;
