+++ source/Tests/Infrastructure/OutgoingMessages/ConfirmRequestChangeAccountingPointCharacteristics/ConfirmRequestChangeAccountingPointCharacteristicsDocumentWriterTests.cs
@@ -21,10 +21,10 @@ using System.Xml.Linq;
 using System.Xml.Schema;
 using Application.Configuration;
 using Application.OutgoingMessages.Common;
+using DocumentValidation;
 using Domain.OutgoingMessages;
 using Infrastructure.Configuration;
 using Infrastructure.Configuration.Serialization;
-using Infrastructure.IncomingMessages.SchemaStore;
 using Infrastructure.OutgoingMessages.Common;
 using Infrastructure.OutgoingMessages.ConfirmRequestChangeAccountingPointCharacteristics;
 using Tests.Infrastructure.OutgoingMessages.Asserts;
