+++ source/Tests/Infrastructure/OutgoingMessages/GenericNotification/GenericNotificationDocumentWriterTests.cs
@@ -21,11 +21,11 @@ using System.Xml.Linq;
 using System.Xml.Schema;
 using Application.Configuration;
 using Application.OutgoingMessages.Common;
+using DocumentValidation;
 using Domain.OutgoingMessages;
 using Domain.OutgoingMessages.GenericNotification;
 using Infrastructure.Configuration;
 using Infrastructure.Configuration.Serialization;
-using Infrastructure.IncomingMessages.SchemaStore;
 using Infrastructure.OutgoingMessages.Common;
 using Infrastructure.OutgoingMessages.GenericNotification;
 using Tests.Infrastructure.OutgoingMessages.Asserts;
