+++ source/Tests/Infrastructure/OutgoingMessages/NotifyAggreagtedMeasureData/NotifyAggregatedMeasureDataDocumentWriterTests.cs
@@ -19,13 +19,13 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Xml.Schema;
 using Application.OutgoingMessages.Common;
+using DocumentValidation;
 using Domain.Actors;
 using Domain.OutgoingMessages;
 using Domain.OutgoingMessages.NotifyAggregatedMeasureData;
 using Domain.SeedWork;
 using Domain.Transactions.Aggregations;
 using Infrastructure.Configuration.Serialization;
-using Infrastructure.IncomingMessages.SchemaStore;
 using Infrastructure.OutgoingMessages.Common;
 using Infrastructure.OutgoingMessages.NotifyAggregatedMeasureData;
 using NodaTime;
