+++ source/Tests/Infrastructure/OutgoingMessages/RejectRequestChangeOfSupplier/JsonDocumentWriterTests.cs
@@ -20,11 +20,11 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using Application.Configuration;
 using Application.OutgoingMessages.Common;
+using DocumentValidation;
 using Domain.OutgoingMessages;
 using Domain.OutgoingMessages.RejectRequestChangeOfSupplier;
 using Infrastructure.Configuration;
 using Infrastructure.Configuration.Serialization;
-using Infrastructure.IncomingMessages.SchemaStore;
 using Infrastructure.OutgoingMessages.Common;
 using Infrastructure.OutgoingMessages.RejectRequestChangeOfSupplier;
 using Json.Schema;
