+++ source/Tests/Infrastructure/OutgoingMessages/RejectRequestChangeOfSupplier/RejectRequestChangeOfSupplierDocumentWriterTests.cs
@@ -21,11 +21,11 @@ using System.Xml.Linq;
 using System.Xml.Schema;
 using Application.Configuration;
 using Application.OutgoingMessages.Common;
+using DocumentValidation;
 using Domain.OutgoingMessages;
 using Domain.OutgoingMessages.RejectRequestChangeOfSupplier;
 using Infrastructure.Configuration;
 using Infrastructure.Configuration.Serialization;
-using Infrastructure.IncomingMessages.SchemaStore;
 using Infrastructure.OutgoingMessages.Common;
 using Infrastructure.OutgoingMessages.RejectRequestChangeOfSupplier;
 using Tests.Infrastructure.OutgoingMessages.Asserts;
