+++ source/Tests/Tests.csproj
@@ -22,7 +22,7 @@ limitations under the License.
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="JsonSchema.Net" Version="3.1.2" />
+        <PackageReference Include="JsonSchema.Net" Version="3.3.2" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4" />
         <PackageReference Include="NJsonSchema" Version="10.7.2" />
         <PackageReference Include="xunit" Version="2.4.1" />
